Added PR fixes
