--- server/lambda/layer/common/nodejs/custom_modules/certificates/read-aws-rds.js
@@ -1,5 +1,0 @@
-const fs = require('fs');
-
-const readAwsRdsCA = () => fs.readFileSync(`${__dirname}/rds-combined-ca-bundle.pem`);
-
-module.exports = readAwsRdsCA;
